Feat/api signup (#75)
* docs: atualização informações readme e .env

* feat(api): criacao rota signup, separação post user
